fastapi_summary_stts/requirements.txt
fastapi
google-generativeai
langchain-google-genai
langchain
pypdf
langchain-community
unstructured
docx2txt
python-magic
jinja2
nest-asyncio
faiss-cpu
tiktoken
networkx
pandas
uvicorn
python-multipart